Moved the colors class to a seperate file
